updated packages util and vanda

+++ share/packages/util/install.bash
@@ -1,15 +1,17 @@
 id="util"
 varname="UTIL"
-version="2013-03-05"
+version="2013-03-07"
 binpath="$id"
 
 install_me () {
 	mkdir -p "$VANDASTUDIO" "$BINDIR" "$FUNCDIR" "$IFACEDIR" "$OUTDIR" "$INDIR" "$PKGDB"
-	echo -e "#!/bin/bash\n"                > "$RCPATH"
-	echo    "DATAPATH=$INDIR"             >> "$RCPATH"
-	echo    "OUTPATH=$OUTDIR"             >> "$RCPATH"
-	echo -e "FUNCDIR=$FUNCDIR\n"          >> "$RCPATH"
-	echo -e "source $FUNCDIR/util.bash\n" >> "$RCPATH"
+	if [ ! -f $RCPATH ]; then
+		echo -e "#!/bin/bash\n"                > "$RCPATH"
+		echo    "DATAPATH=$INDIR"             >> "$RCPATH"
+		echo    "OUTPATH=$OUTDIR"             >> "$RCPATH"
+		echo -e "FUNCDIR=$FUNCDIR\n"          >> "$RCPATH"
+		echo -e "source $FUNCDIR/util.bash\n" >> "$RCPATH"
+	fi
 	cp interfaces/* "$IFACEDIR/."
 	cp -r examples/* "$INDIR/."
 }
